Remove DependencyProvider
