+++ src/main/java/net/fabricmc/loom/extension/LoomGradleExtensionImpl.java
@@ -49,6 +49,8 @@ import net.fabricmc.loom.configuration.InstallerData;
 import net.fabricmc.loom.configuration.LoomDependencyManager;
 import net.fabricmc.loom.configuration.accesswidener.AccessWidenerFile;
 import net.fabricmc.loom.configuration.processors.JarProcessorManager;
+import net.fabricmc.loom.configuration.providers.mappings.MappingsProviderImpl;
+import net.fabricmc.loom.configuration.providers.minecraft.MinecraftProvider;
 
 public class LoomGradleExtensionImpl extends LoomGradleExtensionApiImpl implements LoomGradleExtension {
 	private final Project project;
@@ -64,6 +66,8 @@ public class LoomGradleExtensionImpl extends LoomGradleExtensionApiImpl implemen
 
 	private LoomDependencyManager dependencyManager;
 	private JarProcessorManager jarProcessorManager;
+	private MinecraftProvider minecraftProvider;
+	private MappingsProviderImpl mappingsProvider;
 	private InstallerData installerData;
 
 	public LoomGradleExtensionImpl(Project project, LoomFiles files) {
@@ -116,6 +120,26 @@ public class LoomGradleExtensionImpl extends LoomGradleExtensionApiImpl implemen
 	@Override
 	public JarProcessorManager getJarProcessorManager() {
 		return Objects.requireNonNull(jarProcessorManager, "Cannot get JarProcessorManager before it has been setup");
+	}
+
+	@Override
+	public MinecraftProvider getMinecraftProvider() {
+		return Objects.requireNonNull(minecraftProvider, "Cannot get MinecraftProvider before it has been setup");
+	}
+
+	@Override
+	public void setMinecraftProvider(MinecraftProvider minecraftProvider) {
+		this.minecraftProvider = minecraftProvider;
+	}
+
+	@Override
+	public MappingsProviderImpl getMappingsProvider() {
+		return Objects.requireNonNull(mappingsProvider, "Cannot get MappingsProvider before it has been setup");
+	}
+
+	@Override
+	public void setMappingsProvider(MappingsProviderImpl mappingsProvider) {
+		this.mappingsProvider = mappingsProvider;
 	}
 
 	@Override
